requirements.txt
beautifulsoup4==4.12.2
crochet==2.0.0
Flask==2.0.2
itemadapter==0.4.0
pydash==5.1.0
pymongo==4.1.0
python-dotenv==1.0.0
python_dateutil==2.8.2
random_user_agent==1.0.1
requests==2.25.1
Scrapy==2.6.3
user_agent==0.1.10
